% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{nscore}
\alias{nscore}
\title{Gaussian Transformation (Normal Score)}
\usage{
nscore(x)
}
\arguments{
\item{x}{Numeric vector to transform.}
}
\value{
A list containing:
\item{nscore}{The normalized score vector.}
\item{trn.table}{A data frame mapping the sorted original \code{x} values to the sorted \code{nscore} values.}
}
\description{
Originally written by Ashton Shortridge, May/June, 2008.
}
