% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtr.R
\name{backtr}
\alias{backtr}
\title{Back Transform Function frm Gaussian distribution back to Normal ,
written by Ashton Shortridge, May/June, 2008}
\usage{
backtr(scores, nscore, tails = "none", draw = TRUE)
}
\arguments{
\item{scores}{Gaussian Raster Data}

\item{nscore}{Gaussian Numeric value}

\item{tails}{Extrapolation options on tail value}
}
\value{
'none' No extrapolation; more extreme score values will revert
to the original min and max values.

'equal' Calculate magnitude in std deviations of the scores about
initial data mean. Extrapolation is linear to these deviations.
will be based upon deviations from the mean of the original:

'separate  This calculates a separate sd for values
above and below the mean.
}
\description{
Back Transform Function frm Gaussian distribution back to Normal ,
written by Ashton Shortridge, May/June, 2008
}
