% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtr.R
\name{backtr}
\alias{backtr}
\title{Back Transform Function frm Gaussian distribution back to Normal ,
written by Ashton Shortridge, May/June, 2008}
\usage{
backtr(scores, nscore, tails = "none", draw = TRUE)
}
\arguments{
\item{Back}{transformation}
}
\value{
Dataframe
}
\description{
Back Transform Function frm Gaussian distribution back to Normal ,
written by Ashton Shortridge, May/June, 2008
}
