% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_estimation.R
\name{calc_res}
\alias{calc_res}
\title{Calculate Resources}
\usage{
calc_res(raster_grade, raster_thickness, area, density = 1)
}
\arguments{
\item{raster_grade}{A SpatRaster containing the estimated grade values.}

\item{raster_thickness}{A SpatRaster containing the estimated thickness values.}

\item{area}{An sf polygon outlining the resource calculation boundary.}

\item{density}{Numeric, the bulk density or specific gravity of the material (default = 1.0).}
}
\value{
A list containing the \code{raster} of calculated metal content (tonnage)
and a \code{table} summarizing the area, expected volume, average grade, and total metal content per polygon.
}
\description{
Calculates the volume, grade, and metal content (tonnage) for a resource block.
This calculation is generalized for any commodity, utilizing geometric volume
and grade block models.
}
