% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscore.R
\name{nscore}
\alias{nscore}
\title{Gaussian Transformation, written by Ashton Shortridge, May/June, 2008}
\usage{
nscore(x)
}
\arguments{
\item{x}{Numeric}
}
\description{
Gaussian Transformation, written by Ashton Shortridge, May/June, 2008
}
